+++ packages/auth/src/providers/cognito/apis/resendSignUpCode.ts
@@ -12,11 +12,12 @@ import { AuthDeliveryMedium } from '../../../types';
 import { assertValidationError } from '../../../errors/utils/assertValidationError';
 import { AuthValidationErrorCode } from '../../../errors/types/validation';
 import { ResendSignUpCodeInput, ResendSignUpCodeOutput } from '../types';
-import { getRegion } from '../utils/clients/CognitoIdentityProvider/utils';
-import { resendConfirmationCode } from '../utils/clients/CognitoIdentityProvider';
+import { getRegionFromUserPoolId } from '../../../foundation/parsers';
 import { getAuthUserAgentValue } from '../../../utils';
 import { getUserContextData } from '../utils/userContextData';
 import { ResendConfirmationException } from '../types/errors';
+import { createResendConfirmationCodeClient } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider';
+import { createCognitoUserPoolEndpointResolver } from '../factories';
 
 /**
  * Resend the confirmation code while signing up
@@ -37,7 +38,7 @@ export async function resendSignUpCode(
 	);
 	const authConfig = Amplify.getConfig().Auth?.Cognito;
 	assertTokenProviderConfig(authConfig);
-	const { userPoolClientId, userPoolId } = authConfig;
+	const { userPoolClientId, userPoolId, userPoolEndpoint } = authConfig;
 	const clientMetadata = input.options?.clientMetadata;
 
 	const UserContextData = getUserContextData({
@@ -45,10 +46,14 @@ export async function resendSignUpCode(
 		userPoolId,
 		userPoolClientId,
 	});
-
+	const resendConfirmationCode = createResendConfirmationCodeClient({
+		endpointResolver: createCognitoUserPoolEndpointResolver({
+			endpointOverride: userPoolEndpoint,
+		}),
+	});
 	const { CodeDeliveryDetails } = await resendConfirmationCode(
 		{
-			region: getRegion(authConfig.userPoolId),
+			region: getRegionFromUserPoolId(authConfig.userPoolId),
 			userAgentValue: getAuthUserAgentValue(AuthAction.ResendSignUpCode),
 		},
 		{