+++ scripts/dts-bundler/dts-bundler.config.js
@@ -44,11 +44,14 @@ const authPackageSrcClientsPath = join(
 	'packages',
 	'auth',
 	'src',
-	'providers',
-	'cognito',
-	'utils',
-	'clients',
+	'foundation',
+	'factories',
+	'serviceClients',
+	'cognitoIdentityProvider',
+	'types',
 );
+
+// packages/auth/src/foundation/factories/serviceClients/cognitoIdentityProvider/types/Sdk.ts
 
 /** @type import('dts-bundle-generator/config-schema').BundlerConfig */
 const config = {
@@ -84,8 +87,7 @@ const config = {
 			filePath: './cognito-identity-provider.d.ts',
 			outFile: join(
 				authPackageSrcClientsPath,
-				'CognitoIdentityProvider',
-				'types.ts',
+				'Sdk.ts',
 			),
 			libraries: {
 				inlinedLibraries: ['@aws-sdk/client-cognito-identity-provider'],